# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gupta Enterprises E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register or sync a Firebase user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserBody"
      responses:
        "200":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # PRODUCTS
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cleared cart
  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/apply-coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      summary: Apply coupon to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyCouponBody"
      responses:
        "200":
          description: Cart with coupon applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid coupon

  # ORDERS
...
  /analytics/dashboard:
    get:
      operationId: getDashboardAnalytics
      tags: [analytics]
      summary: Get admin dashboard analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year]
            default: month
      responses:
        "200":
          description: Dashboard analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardAnalytics"
  /analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [analytics]
      summary: Get revenue chart data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"
  /analytics/top-products:
    get:
      operationId: getTopProducts
      tags: [analytics]
      summary: Get top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        firebaseUid:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [USER, ADMIN, DELIVERY_AGENT]
        photoUrl:
          type: string
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/Address"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - createdAt
    Address:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        lat:
          type: number
        lng:
          type: number
        isDefault:
          type: boolean
      required:
        - id
        - street
        - city
        - state
        - pincode
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        actualPrice:
          type: number
        sellingPrice:
          type: number
        discount:
          type: number
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - actualPrice
        - sellingPrice
        - stock
        - images
        - isActive
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
    CartItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - quantity
        - price
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        couponDiscount:
          type: number
        total:
          type: number
        couponCode:
          type: string
        itemCount:
          type: integer
      required:
        - items
        - subtotal
        - discount
        - couponDiscount
        - total
        - itemCount
...
    DashboardAnalytics:
      type: object
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        totalUsers:
          type: integer
        totalProducts:
          type: integer
        revenueGrowth:
          type: number
        ordersGrowth:
          type: number
        usersGrowth:
          type: number
        pendingOrders:
          type: integer
        deliveredOrders:
          type: integer
        cancelledOrders:
          type: integer
        avgOrderValue:
          type: number
      required:
        - totalRevenue
        - totalOrders
        - totalUsers
        - totalProducts
        - pendingOrders
        - deliveredOrders
        - cancelledOrders
        - avgOrderValue
    RevenueDataPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required:
        - date
        - revenue
        - orders
    TopProduct:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        image:
          type: string
        totalSold:
          type: integer
        revenue:
          type: number
      required:
        - productId
        - name
        - totalSold
        - revenue
...
    RegisterUserBody:
      type: object
      properties:
        firebaseUid:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
      required:
        - firebaseUid
        - email
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/Address"
...
    AddToCartBody:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
      required:
        - productId
        - quantity
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity
    ApplyCouponBody:
      type: object
      properties:
        code:
          type: string
      required:
        - code
...