# lib/api-zod/src/generated/types/dashboardAnalytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gupta Enterprises E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardAnalytics {
  totalRevenue: number;
  totalOrders: number;
  totalUsers: number;
  totalProducts: number;
  revenueGrowth?: number;
  ordersGrowth?: number;
  usersGrowth?: number;
  pendingOrders: number;
  deliveredOrders: number;
  cancelledOrders: number;
  avgOrderValue: number;
}