# lib/api-zod/src/generated/types/registerUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gupta Enterprises E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterUserBody {
  firebaseUid: string;
  email: string;
  name?: string;
  phone?: string;
  photoUrl?: string;
}